pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "docugym"
version = "0.1.0"
description = "Local live nature-documentary-style narrator for Gymnasium environments"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
  "ale-py>=0.11,<1",
  "box2d>=2.3,<3",
  "gym>=0.26,<1",
  "gymnasium>=1.2,<2.0",
  "huggingface-sb3>=3.0,<4",
  "httpx>=0.27,<1",
  "numpy>=1.26,<3",
  "pillow>=10,<12",
  "pydantic>=2.7,<3",
  "pydantic-settings>=2.2,<3",
  "pygame>=2.6,<3",
  "pyyaml>=6.0,<7",
  "shimmy>=2.0,<3",
  "stable-baselines3>=2.3,<3",
  "typer>=0.12,<1",
]

[project.scripts]
docugym = "docugym.cli:app"

[dependency-groups]
dev = [
  "black>=24.10,<26",
  "pre-commit>=4.0,<5",
  "pytest>=8.2,<9",
  "ruff>=0.15,<1",
]

[tool.black]
target-version = ['py39', 'py310', 'py311', 'py312']

[tool.ruff]
line-length = 88
target-version = "py312"
fix = true

[tool.ruff.lint]
unfixable = []
select = [
  # pyflakes
  "F",
  # pycodestyle
  "E", "W",
  # flake8-2020
  "YTT",
  # flake8-bugbear
  "B",
  # flake8-quotes
  "Q",
  # pylint
  "PLE", "PLR", "PLW",
  # misc lints
  "PIE",
  # flake8-pyi
  "PYI",
  # tidy imports
  "TID",
  # implicit string concatenation
  "ISC",
  # flake8-type-checking
  "TC",
  # comprehensions
  "C4",
]

ignore = [
  # space before : (needed for how black formats slicing)
  # "E203",  # not yet implemented
  # module level import not at top of file
  "E402",
  # do not assign a lambda expression, use a def
  "E731",
  # line break before binary operator
  # "W503",  # not yet implemented
  # line break after binary operator
  # "W504",  # not yet implemented
  # controversial
  "B006",
  # controversial
  "B007",
  # controversial
  "B008",
  # setattr is used to side-step mypy
  "B009",
  # getattr is used to side-step mypy
  "B010",
  # tests use assert False
  "B011",
  # tests use comparisons but not their returned value
  "B015",
  # false positives
  "B019",
  # Loop control variable overrides iterable it iterates
  "B020",
  # Function definition does not bind loop variable
  "B023",
  # Functions defined inside a loop must not use variables redefined in the loop
  # "B301",  # not yet implemented
  # Only works with python >=3.10
  "B905",
  # Too many arguments to function call
  "PLR0913",
  # Too many returns
  "PLR0911",
  # Too many branches
  "PLR0912",
  # Too many statements
  "PLR0915",
  # Redefined loop name
  "PLW2901",
  # Global statements are discouraged
  "PLW0603",
  # Docstrings should not be included in stubs
  "PYI021",
  # Use typing_extensions.TypeAlias for type aliases
  # "PYI026",  # not yet implemented
  # Use "collections.abc.*" instead of "typing.*" (PEP 585 syntax)
  # "PYI027",  # not yet implemented
  # while int | float can be shortened to float, the former is more explicit
  # "PYI041",  # not yet implemented

  # Additional checks that don't pass yet
  # Useless statement
  "B018",
  # Within an except clause, raise exceptions with ...
  "B904",
  # Magic number
  "PLR2004",
  # Consider `elif` instead of `else` then `if` to remove indentation level
  "PLR5501",
]


[tool.isort]
profile = "black"
combine_as_imports = true
force_grid_wrap = 0
force_sort_within_sections = true

[tool.codespell]
ignore-words-list = "blocs, coo, hist, nd, sav, ser, recuse, nin, timere"
ignore-regex = 'https://([\w/\.])+'


[tool.mypy]
exclude = "data"
# Import discovery
namespace_packages = false
explicit_package_bases = false
ignore_missing_imports = true
follow_imports = "normal"
follow_imports_for_stubs = false
no_site_packages = false
no_silence_site_packages = false
# None and Optional handling
no_implicit_optional = true
strict_optional = true
# Configuring warnings
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
# Suppressing errors
ignore_errors = false
enable_error_code = "ignore-without-code"
# Miscellaneous strictness flags
allow_untyped_globals = false
allow_redefinition = false
local_partial_types = false
implicit_reexport = true
strict_equality = true
# Configuring error messages
show_error_context = false
show_column_numbers = false
show_error_codes = true

[tool.pytest.ini_options]
testpaths = ["tests"]
